Factor commands out of the main file.
