+++ game_data.py
@@ -1,0 +1,18 @@
+from mudserver import MudServer
+
+# structure defining the rooms in the game. Try adding more rooms to the game!
+rooms = {
+    "Tavern": {
+        "description": "You're in a cozy tavern warmed by an open fire.",
+        "exits": {"outside": "Outside"},
+    },
+    "Outside": {
+        "description": "You're standing outside a tavern. It's raining.",
+        "exits": {"inside": "Tavern"},
+    }
+}
+
+players = {}
+
+# start the server
+mud = MudServer()
